Write styles for catalogItem
