+++ src/components/modules/ItemPage/CatalogItem.tsx
@@ -1,7 +1,3 @@
-import CartActiveSvg from 'components/elements/CartActiveSvg/CartActiveSvg';
-import CartInactiveSvg from 'components/elements/CartInactiveSvg/CartInactiveSvg';
-import FavoriteActiveSvg from 'components/elements/FavoriteActiveSvg/FavoriteActiveSvg';
-import FavoriteInactiveSvg from 'components/elements/FavoriteInactiveSvg/FavoriteInactiveSvg';
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { selectCart } from 'store/cart/selectors';
@@ -16,6 +12,7 @@ import ButtonHoverCart from './ButtonHoverCart';
 import CartHoverSvg from 'components/elements/CartHoverSvg/CartHoverSvg';
 import ButtonHoverFavorite from './ButtonHoverFavorite';
 import FavoriteHoverSvg from 'components/elements/FavoriteHoverSvg/FavoriteHoverSvg';
+import styles from 'styles/catalogItem/index.module.scss';
 
 const CatalogItem = ({ item }: CatalogItemType) => {
   const dispatch = useAppDispatch();
@@ -51,10 +48,16 @@ const CatalogItem = ({ item }: CatalogItemType) => {
   };
 
   return (
-    <li className="test__li" key={item.id}>
-      <div className="test__wrapper">
-        <img src={item.titleImage!} alt={item.title} width="302px" height="368px" />
-        <div className="test__wrapper__hover">
+    <li className={styles.catalog__list} key={item.id}>
+      <div className={styles.catalog__list__item}>
+        <img
+          src={item.titleImage!}
+          alt={item.title}
+          width="302px"
+          height="368px"
+          className={styles.catalog__list__item__image}
+        />
+        <div className={styles.catalog__list__wrapper__hover}>
           <button onClick={() => CartItemHandler(item)}>
             <ButtonHoverCart hover={<CartHoverSvg />} isAdded={isAddedToCart} />
           </button>
@@ -64,9 +67,11 @@ const CatalogItem = ({ item }: CatalogItemType) => {
         </div>
       </div>
 
-      <span>Автор: {item.author}</span>
-      <span>Назва: {item.title}</span>
-      <span>{item.price} грн.</span>
+      <div className={styles.catalog__list__item__text}>
+        <span className={styles.catalog__list__item__text__author}>Автор: {item.author}</span>
+        <span className={styles.catalog__list__item__text__title}>Назва: {item.title}</span>
+        <span className={styles.catalog__list__item__text__price}>{item.price} грн.</span>
+      </div>
     </li>
   );
 };
