+++ src/index.scss
@@ -18,39 +18,3 @@ body {
 .main {
   flex: 1;
 }
-
-.test {
-  display: flex;
-  list-style: none;
-  flex-wrap: wrap;
-
-  .test__wrapper {
-    position: relative;
-
-    &__hover {
-      position: absolute;
-      width: 302px;
-      height: 368px;
-      top: 161px;
-      left: 91px;
-
-      button {
-        background: none;
-        border: none;
-        width: 48px;
-        height: 48px;
-        margin-right: 24px;
-      }
-    }
-  }
-
-  &__li {
-    border: 1px solid black;
-    margin-right: 10px;
-    margin-bottom: 10px;
-    padding: 10px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-}
