+++ src/styles/catalogItem/index.module.scss
@@ -1,0 +1,104 @@
+.catalog {
+  display: grid;
+
+  grid-template-rows: 1fr 1fr 1fr 1fr;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  gap: 24px;
+  margin-left: 80px;
+  margin-right: 80px;
+
+  &__list:first-child {
+    grid-row-start: 1;
+    grid-column-start: 1;
+
+    grid-row-end: 3;
+    grid-column-end: 3;
+    padding-bottom: 32px;
+
+    &__item {
+      &__image {
+        width: 628px;
+        height: 859px;
+      }
+    }
+  }
+
+  //&__list:last-child {
+  //  grid-row-start: 3;
+  //  grid-column-start: 3;
+
+  //  grid-row-end: 5;
+  //  grid-column-end: 5;
+  //}
+
+  &__list {
+    border: 1px solid black;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    &__item {
+      background-color: #f5f5f5;
+      width: 100%;
+      font-family: 'Inter';
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      align-items: flex-end;
+
+      &__image {
+        width: 100%;
+      }
+
+      &__text {
+        display: flex;
+        flex-direction: column;
+        justify-content: left;
+        align-items: flex-start;
+
+        &__author {
+          font-weight: 400;
+          font-size: 16px;
+          line-height: 19.36px;
+          color: rgba(149, 149, 149, 1);
+          margin-bottom: 8px;
+        }
+
+        &__title {
+          font-weight: 400;
+          font-size: 20px;
+          line-height: 24.2px;
+          color: rgba(0, 0, 0, 1);
+          margin-bottom: 8px;
+        }
+
+        &__price {
+          font-weight: 600;
+          font-size: 20px;
+          line-height: 24.2px;
+          color: rgba(0, 0, 0, 1);
+        }
+      }
+    }
+
+    &__wrapper {
+      position: relative;
+
+      &__hover {
+        position: absolute;
+        width: 302px;
+        height: 368px;
+        top: 161px;
+        left: 91px;
+
+        button {
+          background: none;
+          border: none;
+          width: 48px;
+          height: 48px;
+          margin-right: 24px;
+        }
+      }
+    }
+  }
+}
